mikrotik connection check

+++ app/Console/Commands/CheckExpiredSubscriptions.php
@@ -49,15 +49,18 @@ class CheckExpiredSubscriptions extends Command
                     $camp_data->mikritikPort
                 );
 
-                // Remove the user from MikroTik
-                $hotspotService->deleteHotspotUser($customer_username);
+                if($hotspotService->isConnected) {
+                    // Remove the user from MikroTik
+                    $hotspotService->deleteHotspotUser($customer_username);
 
-                if(!empty($customer_mac)) {
-                    // Unbind the MAC address from the user
-                    $hotspotService->unbindMacAddressFromUser($customer_mac);
-                }
+                    if(!empty($customer_mac)) {
+                        // Unbind the MAC address from the user
+                        $hotspotService->unbindMacAddressFromUser($customer_mac);
+                    }
 
-                $this->info("Removed expired user: {$customer_username} from camp: {$camp_data->name}");
+                    $this->info("Removed expired user: {$customer_username} from camp: {$camp_data->name}");
+                }//check connection
+
             } else {
                 $this->error("Camp data not found for customer ID: {$customer->id}");
             }
